Moved directory and moved video logic to custom hook
