+++ pages/_app.tsx
@@ -1,6 +1,6 @@
 import "../styles/globals.css";
 import type { AppProps } from "next/app";
-import Header from "../components/Header";
+import Header from "../src/components/Header";
 import { ApolloProvider } from "@apollo/client";
 import client from "../apollo-client";
 import { createPagesBrowserClient } from "@supabase/auth-helpers-nextjs";
@@ -8,8 +8,8 @@ import { SessionContextProvider, Session } from "@supabase/auth-helpers-react";
 import { useState } from "react";
 import { Toaster } from "react-hot-toast";
 import {Provider} from 'react-redux';
-import {store} from "../store";
-import { supabase } from "../components/utils/supabase";
+import store from "../store";
+import { supabase } from "../src/components/utils/supabase";
 export default function App({
   Component,
   pageProps: { ...pageProps },
