+++ pages/index.tsx
@@ -1,10 +1,10 @@
 import Head from "next/head";
-import Header from "../components/Header";
+import Header from "../src/components/Header";
 import { Roboto } from "next/font/google";
 import { Auth, ThemeSupa } from "@supabase/auth-ui-react";
 import { useSession, useSupabaseClient } from "@supabase/auth-helpers-react";
-import Account from "../components/Account";
-import SuggestedVideos from "../components/SuggestedVideos";
+import Account from "../src/components/Account";
+import SuggestedVideos from "../src/components/SuggestedVideos";
 
 const roboto = Roboto({
   weight: "400",
