+++ pages/likedVideo.tsx
@@ -3,7 +3,7 @@ import { useUser } from "@supabase/auth-helpers-react";
 import React from "react";
 import { GET_LIKED_VIDEOS_BY_USER_ID } from "../graphql/queries";
 import styles from "./styles/likedVideo.module.css";
-import VideoIcon from "../components/VideoIcon";
+import VideoIcon from "../src/components/VideoIcon";
 import { LineWobble } from "@uiball/loaders";
 import Link from "next/link";
