+++ src/components/CommentBox.jsx
@@ -1,9 +1,9 @@
 import React, { useState } from "react";
 import Avatar from "./Avatar";
-import { MAKE_COMMENT } from "../graphql/mutations";
-import { GET_VIDEO_BY_ID } from "../graphql/queries";
+import { MAKE_COMMENT } from "../../graphql/mutations";
+import { GET_VIDEO_BY_ID } from "../../graphql/queries";
 import { useMutation } from "@apollo/client";
-import client from "../apollo-client";
+import client from "../../apollo-client";
 import { toast } from "react-hot-toast";
 
 function CommentBox({ video, user }) {
