+++ src/components/styles/LeftHeaderEnd.module.css
