+++ src/types/Subscriber.ts
@@ -1,0 +1,6 @@
+export interface Subscriber {
+  __typename: "Subscribers";
+  id: string;
+  subscribed_to_id: string;
+  user_id: string;
+}
