+++ stepzen.config.json
@@ -1,0 +1,4 @@
+{
+  "endpoint": "api/peeking-whale",
+  "root": "stepzen"
+}
